Create module folders and split the parser
Created new folders for the modules and split the tests from the main
module code.
Split the parser into sub modules
