--- examples/hello.c
@@ -1,6 +1,0 @@
-int x;
-
-int main() {
-   x++;
-   return x;
-}
