--- examples/hello.ll
@@ -1,28 +1,0 @@
-; ModuleID = 'hello.c'
-source_filename = "hello.c"
-target datalayout = "e-m:e-i64:64-f80:128-n8:16:32:64-S128"
-target triple = "x86_64-pc-linux-gnu"
-
-@x = common dso_local local_unnamed_addr global i32 0, align 4
-
-; Function Attrs: nofree norecurse nounwind sspstrong uwtable
-define dso_local i32 @main() local_unnamed_addr #0 {
-  %1 = load i32, i32* @x, align 4, !tbaa !4
-  %2 = add nsw i32 %1, 1
-  store i32 %2, i32* @x, align 4, !tbaa !4
-  ret i32 %2
-}
-
-attributes #0 = { nofree norecurse nounwind sspstrong uwtable "correctly-rounded-divide-sqrt-fp-math"="false" "disable-tail-calls"="false" "less-precise-fpmad"="false" "min-legal-vector-width"="0" "no-frame-pointer-elim"="false" "no-infs-fp-math"="false" "no-jump-tables"="false" "no-nans-fp-math"="false" "no-signed-zeros-fp-math"="false" "no-trapping-math"="false" "stack-protector-buffer-size"="8" "target-cpu"="x86-64" "target-features"="+cx8,+fxsr,+mmx,+sse,+sse2,+x87" "unsafe-fp-math"="false" "use-soft-float"="false" }
-
-!llvm.module.flags = !{!0, !1, !2}
-!llvm.ident = !{!3}
-
-!0 = !{i32 1, !"wchar_size", i32 4}
-!1 = !{i32 7, !"PIC Level", i32 2}
-!2 = !{i32 7, !"PIE Level", i32 2}
-!3 = !{!"clang version 9.0.0 (tags/RELEASE_900/final)"}
-!4 = !{!5, !5, i64 0}
-!5 = !{!"int", !6, i64 0}
-!6 = !{!"omnipotent char", !7, i64 0}
-!7 = !{!"Simple C/C++ TBAA"}
