--- examples/struct.ll
@@ -1,27 +1,0 @@
-; ModuleID = 'struct.c'
-source_filename = "struct.c"
-target datalayout = "e-m:e-i64:64-f80:128-n8:16:32:64-S128"
-target triple = "x86_64-pc-linux-gnu"
-
-%struct.refs_interface = type { i32, i32 }
-
-@refs_instance = common dso_local global %struct.refs_interface zeroinitializer, align 4
-
-; Function Attrs: noinline nounwind optnone sspstrong uwtable
-define dso_local i32 @main() #0 {
-  %1 = alloca i32, align 4
-  store i32 0, i32* %1, align 4
-  %2 = load i32, i32* getelementptr inbounds (%struct.refs_interface, %struct.refs_interface* @refs_instance, i32 0, i32 0), align 4
-  %3 = add nsw i32 %2, 1337
-  ret i32 %3
-}
-
-attributes #0 = { noinline nounwind optnone sspstrong uwtable "correctly-rounded-divide-sqrt-fp-math"="false" "disable-tail-calls"="false" "less-precise-fpmad"="false" "min-legal-vector-width"="0" "no-frame-pointer-elim"="true" "no-frame-pointer-elim-non-leaf" "no-infs-fp-math"="false" "no-jump-tables"="false" "no-nans-fp-math"="false" "no-signed-zeros-fp-math"="false" "no-trapping-math"="false" "stack-protector-buffer-size"="8" "target-cpu"="x86-64" "target-features"="+cx8,+fxsr,+mmx,+sse,+sse2,+x87" "unsafe-fp-math"="false" "use-soft-float"="false" }
-
-!llvm.module.flags = !{!0, !1, !2}
-!llvm.ident = !{!3}
-
-!0 = !{i32 1, !"wchar_size", i32 4}
-!1 = !{i32 7, !"PIC Level", i32 2}
-!2 = !{i32 7, !"PIE Level", i32 2}
-!3 = !{!"clang version 9.0.0 (tags/RELEASE_900/final)"}
