+++ src/ast.rs
@@ -1,5 +1,3 @@
-pub type Return<T> = Option<T>;
-
 #[derive(Debug, PartialEq)]
 pub struct Program {
     pub name: String,
