+++ src/parser/control/tests.rs
@@ -1,0 +1,140 @@
+use crate::lexer;
+use crate::parser::parse;
+use pretty_assertions::*;
+
+
+#[test]
+fn if_statement() {
+    let lexer = lexer::lex(
+        "
+        PROGRAM exp 
+        IF TRUE THEN
+        END_IF
+        END_PROGRAM
+        ",
+    );
+    let result = parse(lexer).unwrap();
+
+    let prg = &result.units[0];
+    let statement = &prg.statements[0];
+
+    let ast_string = format!("{:#?}", statement);
+    let expected_ast = r#"IfStatement {
+    blocks: [
+        ConditionalBlock {
+            condition: LiteralBool {
+                value: true,
+            },
+            body: [],
+        },
+    ],
+    else_block: [],
+}"#;
+    assert_eq!(ast_string, expected_ast);
+}
+
+#[test]
+fn if_else_statement_with_expressions() {
+    let lexer = lexer::lex(
+        "
+        PROGRAM exp 
+        IF TRUE THEN
+            x;
+        ELSE
+            y;
+        END_IF
+        END_PROGRAM
+        ",
+    );
+    let result = parse(lexer).unwrap();
+
+    let prg = &result.units[0];
+    let statement = &prg.statements[0];
+
+    let ast_string = format!("{:#?}", statement);
+    let expected_ast = r#"IfStatement {
+    blocks: [
+        ConditionalBlock {
+            condition: LiteralBool {
+                value: true,
+            },
+            body: [
+                Reference {
+                    name: "x",
+                },
+            ],
+        },
+    ],
+    else_block: [
+        Reference {
+            name: "y",
+        },
+    ],
+}"#;
+    assert_eq!(ast_string, expected_ast);
+}
+
+#[test]
+fn if_elsif_elsif_else_statement_with_expressions() {
+    let lexer = lexer::lex(
+        "
+        PROGRAM exp 
+        IF TRUE THEN
+            x;
+        ELSIF y THEN
+            z;
+        ELSIF w THEN
+            v;
+        ELSE
+            u;
+        END_IF
+        END_PROGRAM
+        ",
+    );
+    let result = parse(lexer).unwrap();
+
+    let prg = &result.units[0];
+    let statement = &prg.statements[0];
+
+    let ast_string = format!("{:#?}", statement);
+    let expected_ast = r#"IfStatement {
+    blocks: [
+        ConditionalBlock {
+            condition: LiteralBool {
+                value: true,
+            },
+            body: [
+                Reference {
+                    name: "x",
+                },
+            ],
+        },
+        ConditionalBlock {
+            condition: Reference {
+                name: "y",
+            },
+            body: [
+                Reference {
+                    name: "z",
+                },
+            ],
+        },
+        ConditionalBlock {
+            condition: Reference {
+                name: "w",
+            },
+            body: [
+                Reference {
+                    name: "v",
+                },
+            ],
+        },
+    ],
+    else_block: [
+        Reference {
+            name: "u",
+        },
+    ],
+}"#;
+    assert_eq!(ast_string, expected_ast);
+}
